fix: resolve STRING type missing ExtraTypeInfo and uint16_t overflow
- convertFromString("STRING") now returns DataType::Varchar() instead of
  DataType(DataTypeId::kVarchar), ensuring StringTypeInfo is attached with
  default max_length=256
- Changed uint16_t to size_t in pb_utils.cc to prevent overflow when
  VARCHAR_DEFAULT_LENGTH (65536) is assigned via protobuf conversion
- Removed debug logging from project_utils.cc

+++ src/utils/pb_utils.cc
@@ -158,7 +158,7 @@ bool string_type_to_property_type(const common::String& string_type,
                                   DataType& out_type) {
   switch (string_type.item_case()) {
   case common::String::kVarChar: {
-    uint16_t max_length = STRING_DEFAULT_MAX_LENGTH;
+    size_t max_length = STRING_DEFAULT_MAX_LENGTH;
     if (string_type.has_var_char()) {
       auto str_info = string_type.var_char();
       if (str_info.max_length() > 0) {
